+++ sol-did/programs/sol-did/src/errors.rs
@@ -6,12 +6,12 @@ pub enum DidSolError {
     VmFragmentNotFound,
     #[msg("Given VM fragment is already in use")]
     VmFragmentAlreadyInUse,
-    #[msg("Cannot add a verification method with OwnershipProof flag")]
-    VmOwnershipOnAdd,
+    #[msg("Cannot add a verification method with guarded flag (OwnershipProof or Protected)")]
+    VmGuardedFlagOnAdd,
     #[msg("Removing the last verification method would lead to a lockout")]
     VmCannotRemoveLastAuthority,
     #[msg("Service already exists in current service list")]
-    ServiceFragmentAlreadyInUse, // TODO: Fragments should be globally unique.
+    ServiceFragmentAlreadyInUse,
     #[msg("Service doesn't exists in current service list")]
     ServiceFragmentNotFound,
     #[msg("Invalid other controllers. Invalid DID format or did:sol:<did>")]
@@ -28,4 +28,6 @@ pub enum DidSolError {
     ErrorValidatingSecp256k1Signature,
     #[msg("Wrong Authority for given DID")]
     WrongAuthorityForDid,
+    #[msg("Cannot remove a protected verification method. You need to first remove the Protected Verification Method Flag in order for this operation to succeed")]
+    VmCannotRemoveProtected,
 }